requirements.txt
absl-py==0.8.1
altair==3.2.0
argh==0.26.2
astor==0.8.0
atomicwrites==1.3.0
attrs==19.3.0
base58==1.0.3
blinker==1.4
boto3==1.10.16
botocore==1.13.16
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
cloudpickle==1.2.2
colorama==0.4.1
configobj==5.0.6
cycler==0.10.0
dask==2.7.0
decorator==4.4.1
distributed==2.7.0
docutils==0.15.2
entrypoints==0.3
enum-compat==0.0.3
filelock==3.0.12
funcsigs==1.0.2
future==0.18.2
gast==0.2.2
google-auth==1.7.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.25.0
h5py==2.10.0
HeapDict==1.0.1
idna==2.8
imageio==2.6.1
importlib-metadata==0.23
Jinja2==2.10.3
jmespath==0.9.4
joblib==0.14.0
jsonschema==3.1.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
llvmlite==0.30.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mlxtend==0.17.0
more-itertools==7.2.0
msalign==0.1.2
msgpack==0.6.2
networkx==2.4
numba==0.46.0
numexpr==2.7.0
numpy==1.17.4
oauthlib==3.1.0
opt-einsum==3.1.0
packaging==19.2
pandas==0.25.3
parmap==1.5.2
pathlib==1.0.1
pathtools==0.1.2
patsy==0.5.1
Pillow==6.2.1
pkg-resources==0.0.0
plotly==4.3.0
pluggy==0.13.0
protobuf==3.10.0
psutil==5.6.5
py==1.8.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
pyparsing==2.4.5
pyrsistent==0.15.5
pytest==5.2.2
python-dateutil==2.8.0
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.1.2
ray==0.7.6
redis==3.3.11
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.0
s3transfer==0.2.1
scikit-image==0.16.2
scikit-learn==0.21.3
scipy==1.3.2
seaborn==0.9.0
six==1.13.0
sortedcontainers==2.1.0
statsmodels==0.10.1
streamlit==0.50.2
tables==3.6.1
tblib==1.5.0
tensorboard==1.15.0
tensorflow-estimator==1.15.0
termcolor==1.1.0
tf-nightly-gpu==1.15.0.dev20190730
toml==0.10.0
toolz==0.10.0
tornado==5.1.1
tqdm==4.38.0
trackpy==0.4.2
tsfresh==0.12.0
tzlocal==2.0.0
urllib3==1.25.7
validators==0.14.0
watchdog==0.9.0
wcwidth==0.1.7
Werkzeug==0.16.0
wrapt==1.11.2
zict==1.0.0
zipp==0.6.0
